requirements.txt
# Core dependencies
rdflib>=7.0.0
pyshacl>=0.25.0
anthropic>=0.7.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.0
Pillow>=10.0.0

# Optional OCR support
pytesseract>=0.3.10

# JSON/YAML support
pyyaml>=6.0.0

# Database schema support (optional)
sqlalchemy>=2.0.0

# CLI support
click>=8.1.0
rich>=13.0.0

# Testing (dev only)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality (dev only)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Documentation (dev only)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
